+++ .github/CHANGELOG.md
@@ -177,6 +177,7 @@
 1. [HYD] Fix nose wheel position during pushback - @Crocket63
 1. [MCDU] Improved Remote MCDU communication by reducing update frequency - @frankkopp (Frank Kopp)
 1. [FLIGHTMODEL] Additional flight model improvements to better match real airplane performance data - @donstim (donbikes#4084)
+1. [FMGC] Re-implement altitude and speed constraints - @tracernz (Mike)
 
 ## 0.7.0